pyproject.toml
[project]
name = "churnguard-analytics"
version = "0.1.0"
description = "ChurnGuard analytics platform"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.74.0",
    "ipykernel>=7.2.0",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "streamlit>=1.57.0",
    "sweetviz>=2.3.3",
    "xgboost>=3.2.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "nbconvert>=7.17.1",
]
